Renaming package files/namespace and moving more functions to it
