+++ dev/Deployment/Deployment.vcxitems
@@ -28,7 +28,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)DeploymentActivityContext.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)PackageDefinitions.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Deployer.h" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)PackagePathUtilities.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)PackageUtilities.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)PackageRegistrar.h" />
   </ItemGroup>
   <ItemGroup>
@@ -38,7 +38,7 @@
     <ClCompile Include="$(MSBuildThisFileDirectory)DeploymentResult.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)DeploymentActivityContext.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)Deployer.cpp" />
-    <ClCompile Include="$(MSBuildThisFileDirectory)PackagePathUtilities.cpp" />
+    <ClCompile Include="$(MSBuildThisFileDirectory)PackageUtilities.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)PackageRegistrar.cpp" />
   </ItemGroup>
   <ItemGroup>
