+++ test/DeploymentUnitTests/DeploymentUnitTests.vcxproj
@@ -216,7 +216,7 @@
     <ClCompile Include="$(RepoRoot)\dev\Deployment\PackageRegistrar.cpp" />
     <ClCompile Include="$(RepoRoot)\dev\Deployment\DeploymentActivityContext.cpp" />
     <ClCompile Include="$(RepoRoot)\dev\Deployment\Deployer.cpp" />
-    <ClCompile Include="$(RepoRoot)\dev\Deployment\PackagePathUtilities.cpp" />
+    <ClCompile Include="$(RepoRoot)\dev\Deployment\PackageUtilities.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
